# Reduce actor submission python overhead