--- a/python/ray/includes/libcoreworker.pxi
+++ b/python/ray/includes/libcoreworker.pxi
@@ -35,8 +35,7 @@ cdef class ProfileEvent:
         elif self.extra_data is not None:
             extra_data = self.extra_data
 
-        if extra_data:
-            self.inner.get().SetExtraData(json.dumps(extra_data).encode("ascii"))
+        self.inner.get().SetExtraData(json.dumps(extra_data).encode("ascii") if extra_data else b"{}")
 
         # Deleting the CProfileEvent will add it to a queue to be pushed to
         # the driver.