--- a/python/ray/local_mode_manager.py
+++ b/python/ray/local_mode_manager.py
@@ -29,8 +29,7 @@ class LocalModeManager(object):
     def __init__(self):
         """Initialize a LocalModeManager."""
 
-    def execute(self, function, function_name, args, kwargs,
-                num_return_vals):
+    def execute(self, function, function_name, args, kwargs, num_return_vals):
         """Synchronously executes a "remote" function or actor method.
 
         Stores results directly in the generated and returned