--- a/src/ray/core_worker/profiling.h
+++ b/src/ray/core_worker/profiling.h
@@ -2,10 +2,10 @@
 #define RAY_CORE_WORKER_PROFILING_H
 
 #include "absl/synchronization/mutex.h"
+#include "absl/time/clock.h"
 
 #include "ray/core_worker/context.h"
 #include "ray/gcs/redis_gcs_client.h"
-#include "ray/util/util.h"
 
 namespace ray {
 
@@ -44,7 +44,7 @@ class ProfileEvent {
   ProfileEvent(const std::shared_ptr<Profiler> profiler, const std::string &event_type);
 
   ~ProfileEvent() {
-    rpc_event_.set_end_time(current_sys_time_seconds());
+    rpc_event_.set_end_time(absl::GetCurrentTimeNanos() / 1e9);
     profiler_->AddEvent(rpc_event_);
   }