--- a/src/ray/util/util.h
+++ b/src/ray/util/util.h
@@ -10,6 +10,8 @@
 #include <string>
 #include <unordered_map>
 
+#include "absl/time/clock.h"
+
 #include "ray/common/status.h"
 
 /// Return the number of milliseconds since the steady clock epoch. NOTE: The
@@ -43,8 +45,8 @@ inline int64_t current_sys_time_us() {
 }
 
 inline double current_sys_time_seconds() {
-  int64_t microseconds_in_seconds = 1000000;
-  return static_cast<double>(current_sys_time_us()) / microseconds_in_seconds;
+  int64_t nanoseconds_in_seconds = 1000000000;
+  return static_cast<double>(absl::GetCurrentTimeNanos()) / nanoseconds_in_seconds;
 }
 
 inline ray::Status boost_to_ray_status(const boost::system::error_code &error) {