@pcmoritz 2019-10-23
Latest numbers with this PR (`core_worker.get().SubmitActorTask` commented out):

After:

```
In [3]: @ray.remote 
   ...: class C: 
   ...:     def f(self): 
   ...:         return 42 
   ...:                                                                                                                                                                              

In [4]: c = C.remote()                                                                                                                                                               

In [5]: %timeit c.f.remote()                                                                                                                                                         
20.5 µs ± 1.29 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

Before:

```
In [8]: %timeit c.f.remote()
69.6 µs ± 5.05 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

There are some more optimizations from #4522 that can be ported now that the core worker execution interface and the kwargs PR are merged (before there would have been a bunch of conflicts). Will do that in a followup PR.