@pcmoritz 2019-10-23
Jenkins failure was introduced by https://github.com/ray-project/ray/pull/5783